\name{diffeR-package}
\alias{diffeR-package}
\alias{diffeR}
\docType{package}
\title{
Difference Metrics for Comparing Pairs of Maps
}
\description{
Difference metrics for comparing pairs of maps representing real or categorical variables at original and multiple resolutions.}
\details{
\tabular{ll}{
Package: \tab composite\cr
Type: \tab Package\cr
Version: \tab 0.0-2\cr
Date: \tab 2014-11-28\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Robert Gilmore Pontius Jr. <rpontius@clarku.edu>, Alí Santacruz <asantacruzdelgado@clarku.edu>

Maintainer: Alí Santacruz <amsantac@unal.edu.co>
}
\keyword{ package }
\keyword{ spatial }
\seealso{
\code{\link{composite}} 
}
