\name{diffeR-package}
\alias{diffeR}
\alias{diffeR-package}
\docType{package}
\title{
Metrics of Difference for Comparing Pairs of Maps or Pairs of Variables
}
\description{
Metrics of difference for comparing pairs of variables or pairs of maps representing real or categorical variables at original and multiple resolutions.}
\details{
\tabular{ll}{
Package: \tab diffeR\cr
Type: \tab Package\cr
Version: \tab 0.0-8\cr
Date: \tab 2023-08-21\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
BugReports: \tab https://github.com/amsantac/diffeR/issues\cr 
}
}
\author{
Robert Gilmore Pontius Jr. <rpontius@clarku.edu>, Ali Santacruz <amsantac@unal.edu.co>

Maintainer: Ali Santacruz <amsantac@unal.edu.co>
}
\keyword{ package }
\keyword{ spatial }
